Improved directory, cache, remove account

+++ iOSClient/Utility/NCUtility.swift
@@ -375,26 +375,22 @@ class NCUtility: NSObject {
         return false
     }
     
-    @objc func removeAccountOnDBKeychain(_ account: String?, onlyLocalFile: Bool) {
+    @objc func removeAccount(_ account: String?, removeKeychain: Bool) {
         
-        if onlyLocalFile {
-            NCManageDatabase.sharedInstance.clearTable(tableLocalFile.self, account: account)
-        } else {
-            NCManageDatabase.sharedInstance.clearTable(tableAccount.self, account: account)
-            NCManageDatabase.sharedInstance.clearTable(tableActivity.self, account: account)
-            NCManageDatabase.sharedInstance.clearTable(tableActivitySubjectRich.self, account: account)
-            NCManageDatabase.sharedInstance.clearTable(tableCapabilities.self, account: account)
-            NCManageDatabase.sharedInstance.clearTable(tableDirectory.self, account: account)
-            NCManageDatabase.sharedInstance.clearTable(tableE2eEncryption.self, account: account)
-            NCManageDatabase.sharedInstance.clearTable(tableExternalSites.self, account: account)
-            NCManageDatabase.sharedInstance.clearTable(tableLocalFile.self, account: account)
-            NCManageDatabase.sharedInstance.clearTable(tableMetadata.self, account: account)
-            NCManageDatabase.sharedInstance.clearTable(tableMedia.self, account: account)
-            NCManageDatabase.sharedInstance.clearTable(tablePhotoLibrary.self, account: account)
-            NCManageDatabase.sharedInstance.clearTable(tableShare.self, account: account)
-        }
+        NCManageDatabase.sharedInstance.clearTable(tableAccount.self, account: account)
+        NCManageDatabase.sharedInstance.clearTable(tableActivity.self, account: account)
+        NCManageDatabase.sharedInstance.clearTable(tableActivitySubjectRich.self, account: account)
+        NCManageDatabase.sharedInstance.clearTable(tableCapabilities.self, account: account)
+        NCManageDatabase.sharedInstance.clearTable(tableDirectory.self, account: account)
+        NCManageDatabase.sharedInstance.clearTable(tableE2eEncryption.self, account: account)
+        NCManageDatabase.sharedInstance.clearTable(tableExternalSites.self, account: account)
+        NCManageDatabase.sharedInstance.clearTable(tableLocalFile.self, account: account)
+        NCManageDatabase.sharedInstance.clearTable(tableMetadata.self, account: account)
+        NCManageDatabase.sharedInstance.clearTable(tableMedia.self, account: account)
+        NCManageDatabase.sharedInstance.clearTable(tablePhotoLibrary.self, account: account)
+        NCManageDatabase.sharedInstance.clearTable(tableShare.self, account: account)
         
-        if (account != nil ) {
+        if (account != nil && removeKeychain) {
             CCUtility.clearAllKeysEnd(toEnd: account)
             CCUtility.clearAllKeysPushNotification(account)
             CCUtility.setCertificateError(account, error: false)
@@ -402,7 +398,7 @@ class NCUtility: NSObject {
         }
     }
     
-    @objc func removeAllSettings() {
+    @objc func removeAllSettings(removeKeychain: Bool) {
         
         URLCache.shared.memoryCapacity = 0
         URLCache.shared.diskCapacity = 0
@@ -417,7 +413,9 @@ class NCUtility: NSObject {
         
         CCUtility.createDirectoryStandard()
         
-        CCUtility.deleteAllChainStore()
+        if removeKeychain {
+            CCUtility.deleteAllChainStore()
+        }
     }
 }
